safe extract for all tar files
